requirements.txt
# Links for torch and dgl
-f https://download.pytorch.org/whl/torch_stable.html
-f https://data.dgl.ai/wheels/repo.html

# Safe version update - should not impact model performance.
tqdm==4.63.0
loguru==0.6.0
requests==2.27.1
pandas==1.5.0
beautifulsoup4==4.10.0
lxml==4.8.0
SPARQLWrapper==2.0.0
seaborn==0.13.2

# Unsafe version update - could impact model performance.
networkx==2.7.1
transformers==4.17.0
scikit-learn==1.3.2
torch==1.9.1+cu111
torchvision==0.10.1+cu111
torchaudio==0.9.1
dgl-cu111==0.8.1
sentence-transformers==0.4.1.2
parameter-sherpa==1.0.6
scipy==1.8.0
tensorboard==2.12.0

# requires pybind11==2.13.1 beforehand
graph-walker==1.0.6
tbparse==0.0.9